Rejigged the B-field color key used in mhdcuts figs
+ mhdcuts-Bkey.py - make key bigger 128x128 and do some variants. Most
are not used, except for the ones called "sat" and "val". In the
"sat", we have the saturation falling off towards the middle of the
disk, corresponding to increasingly prominent out-of-plane component
of the field. But unlike in the original version, we also reduce the
value slightly in the middle (to 0.8), so it doesn't look so washed
out there. In "val", the value is reduced towrads the center,
representing reduced B field.

+ mhdcuts-B30krum-stitchup-nolabels.py and krumx too - use the "sat"
and "val" variants of the color key. Put them at the top of the top
panels, instead of between top and bottom panels.